refactor: keep sealos sdk integration under integrations
