+++ integrations/sealos/desktop-sdk/get-sealos-session.ts
