+++ integrations/sealos/desktop-sdk/types.ts
