+++ integrations/sealos/index.ts
@@ -1,0 +1,5 @@
+export {
+  detectSealosIframe,
+  getSealosSession,
+} from './desktop-sdk'
+export type { SealosAuthSession, SealosUserInfo } from './desktop-sdk'
